requirementsT.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
addict=2.2.1=pypi_0
blas=1.0=mkl
ca-certificates=2019.10.16=0
certifi=2019.9.11=py37_0
cffi=1.12.3=py37h2e261b9_0
chardet=3.0.4=pypi_0
cudatoolkit=10.0.130=0
cudnn=7.6.0=cuda10.0_0
cycler=0.10.0=pypi_0
cython=0.29.13=pypi_0
decorator=4.4.0=pypi_0
freetype=2.9.1=h8a8886c_1
icu=58.2=h9c2bf20_1
idna=2.8=pypi_0
imagecorruptions=1.0.0=pypi_0
imageio=2.5.0=pypi_0
intel-openmp=2019.4=243
jpeg=9b=0
kiwisolver=1.1.0=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libxml2=2.9.9=hea5a465_1
libxslt=1.1.33=h7d1a2b0_0
lxml=4.4.1=py37hefd8a0e_0
matplotlib=3.1.1=pypi_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.14=py37ha843d7b_0
mkl_random=1.0.2=py37hd81dba3_0
mmcv=0.2.13=pypi_0
mmdet=1.0rc0+unknown=dev_0
ncurses=6.1=he6710b0_1
networkx=2.3=pypi_0
ninja=1.7.2=0
numpy=1.16.4=py37h7e9f1db_0
numpy-base=1.16.4=py37hde5b4d6_0
olefile=0.46=py37_0
opencv-python=4.1.1.26=pypi_0
openssl=1.1.1d=h7b6447c_3
pillow=6.1.0=py37h34e0f95_0
pip=19.2.2=py37_0
pycocotools=2.0.0=pypi_0
pycparser=2.19=py37_0
pyparsing=2.4.2=pypi_0
python=3.7.4=h265db76_1
python-dateutil=2.8.0=pypi_0
pytorch=1.1.0=cuda100py37he554f03_0
pywavelets=1.0.3=pypi_0
pyyaml=5.1.2=pypi_0
readline=7.0=h7b6447c_5
requests=2.22.0=pypi_0
scikit-image=0.15.0=pypi_0
scipy=1.3.1=pypi_0
setuptools=41.0.1=py37_0
six=1.12.0=py37_0
sqlite=3.29.0=h7b6447c_0
terminaltables=3.1.0=pypi_0
tk=8.6.8=hbc83047_0
torchvision=0.4.0=pypi_0
urllib3=1.25.3=pypi_0
wheel=0.33.4=py37_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=0
zstd=1.3.7=h0b5b093_0
